pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safeprobe"
version = "1.0.0"
description = "An open-source Python toolkit for evaluating the safety alignment of LLMs during inference, focusing on query-access attack vectors."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Milton Pedro Pagliuso Neto", email = "milton.neto@edu.udesc.br"}
]
keywords = [
    "llm", "safety-alignment", "adversarial-machine-learning",
    "red-teaming", "jailbreaking", "ai-safety", "nist-taxonomy"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "litellm>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "tiktoken>=0.5.0",
    "psutil>=5.9.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black",
    "flake8",
]
ollama = ["ollama>=0.1.0"]
google = ["google-genai>=0.3.0"]
pair = ["fschat>=0.2.0", "wandb>=0.15.0"]
full = [
    "ollama>=0.1.0",
    "google-genai>=0.3.0",
    "fschat>=0.2.0",
    "wandb>=0.15.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/tompagli/safeprobe"
Documentation = "https://github.com/tompagli/safeprobe/wiki"
Repository = "https://github.com/tompagli/safeprobe"

[project.scripts]
safeprobe = "safeprobe.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
safeprobe = [
    "attacks/promptmap/rules/**/*.yaml",
    "attacks/cipherchat/data/*",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
